loads images and displays one
I've created a Picture class to abstract image scaling and store info. I also improved the logging system, though I do not consider it complete. How I'm currently thinking it'll work is that there will be functions called 'renderers,' which render a transition at a certian amount of progress. Currently, the only one is the 'static' renderer, which is used when no transition is taking place.