display wrapper
There is now a Picture object to represent the display which is passed to renderers instead of the actual display surface. This is so that renderers can have access to all the measurements you get on a Picture for the display. There are also more measurements available on Picture objects. Pictures now have a 'scale' method which smoothly scales the surface if possible and normally scales it otherwise. The slideshow also now supports images below 24-bit and doesn't crash when it tries to load them. I also have some new sample images which are pictures of Nintendo games, except the sample BMP, which is a pumpkin meant to only be displayed in October. Currently, date filtering is not implemented, but I intend to add it relatively soon.